fix: fill src attribute from data-src on images
To improve support for NoScript users, we need a way to handle image
with data-src attributes which is a common way for many websites to do
image lazy loading.
This patch loop over all images and fix the src attribute after all
processing has been done.
Originally I intended to fix the src attribute on the cloneImages
function but found out that some images were ignored for some reasons.